requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
harperdb==0.4.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
zope.event==4.5.0
